+++ portal/apps/core/views/edition.py
@@ -39,7 +39,8 @@ def edition_list(request):
 @to_response
 def edition_list_ajax(request, year, month):
     editions = Edition.objects.all().order_by('-date_published').filter(
-        date_published__year=year, date_published__month=month)
+        date_published__year=year, date_published__month=month
+    )
     cal = EditionCalendar(editions, settings.FIRST_DAY_OF_WEEK, settings.LOCALE_NAME).formatmonth(year, month)
     return 'list_ajax.html', {'calendar': mark_safe(cal)}
 
@@ -62,12 +63,16 @@ def is_valid_download_link(request):
 
 @never_cache
 def get_download_validation_url(edition, user):
-    url = reverse('edition_download', args=(), kwargs={
-        'year': edition.date_published.year,
-        'month': edition.date_published.month,
-        'day': edition.date_published.day})
-    return '%s?token=%i-%s' % (
-        url, user.id, download_token_generator.make_token(user))
+    url = reverse(
+        'edition_download',
+        args=(),
+        kwargs={
+            'year': edition.date_published.year,
+            'month': edition.date_published.month,
+            'day': edition.date_published.day,
+        },
+    )
+    return '%s?token=%i-%s' % (url, user.id, download_token_generator.make_token(user))
 
 
 @never_cache
@@ -81,15 +86,18 @@ def edition_download(request, publication_slug, year, month, day, filename):
     edition = get_object_or_404(Edition, publication__slug=publication_slug, date_published=date_published)
     downloadable_editions = getattr(settings, 'CORE_PUBLICATIONS_EDITION_DOWNLOAD', ())
     if request.user.is_staff or (
-            publication_slug == settings.DEFAULT_PUB and request.user.subscriber.is_subscriber() and
-            request.user.subscriber.pdf
-        ) or (
-            publication_slug in downloadable_editions and request.user.subscriber.is_subscriber(publication_slug) and
-            getattr(request.user.subscriber, publication_slug + '_pdf', False)):
+        publication_slug == settings.DEFAULT_PUB and request.user.subscriber.is_subscriber()
+        and request.user.subscriber.pdf
+    ) or (
+        publication_slug in downloadable_editions and request.user.subscriber.is_subscriber(publication_slug)
+        and getattr(request.user.subscriber, publication_slug + '_pdf', False)
+    ):
         if request.user.is_superuser or (date.today() - timedelta(70) < date_published):
             return (
-                edition if edition.pdf.name.endswith(filename) else
-                get_object_or_404(Supplement, edition=edition, pdf__endswith=filename)).download(request)
+                edition if edition.pdf.name.endswith(filename) else get_object_or_404(
+                    Supplement, edition=edition, pdf__endswith=filename
+                )
+            ).download(request)
 
     return HttpResponseForbidden()
 
@@ -97,9 +105,9 @@ def edition_download(request, publication_slug, year, month, day, filename):
 def rawpic_cover(request):
     edition = get_current_edition()
     if not edition:
-        response = HttpResponse(
-            "No hay foto disponible!", mimetype="text/plain")
+        raise Http404
     else:
-        rawpic = open(edition.cover.path)
-        response = HttpResponse(rawpic.read(), mimetype="image/jpeg")
-    return response
+        try:
+            return HttpResponse(open(edition.cover.path).read(), content_type="image/jpeg")
+        except IOError:
+            raise Http404